+++ docs/changelog.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added:
 - Get missing stats and files in bedfiles
+- Config analyzer for quick analysis if the config is correct
 
 
 ## [0.10.2] - 2025-01-09